auth_project/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
mysqlclient